Check version sources based on whether verison has been set
Update logic for reading sources for obtaining version to read from next source explicitly if version has not been set.
This handles situations where a source exists (e.g. TOML file, terraform files), but a version constraint cannot be found from the source.
This avoids duplicate checks for existance of sources (such as Terragrunt file), as this was performed in GetParamters and also in the methods to read the config from the source.

Issue #420