feat: Support falling back to Terraform for pre-1.6 releases
This adds an option to automatically fall back to using Terraform
when the version required from a terraform module is older than 1.6.0.
This will aid teams working across multiple repositories or monorepos
that have modules that haven't yet been upgraded.

I only implemented the fallback for versiontf handling, as I expect
that is the place it will most be needed.